db-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
motor==3.3.1
pymongo==4.5.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
loguru==0.7.2
certifi>=2023.7.22
pytest-asyncio>=0.21.1
httpx>=0.25.2 